config: introduce AtomicString for config items
